rename number range and update imports
